pyproject.toml
[tool.poetry]
name = "resume-collector"
version = "0.1.0"
description = ""
authors = ["them1n9 <iamanidiotm@gmail.com>"]
readme = "README.md"

packages = [
    { include = "./recruiter/nlqs" },
    { include = "./recruiter/scripts" },
    { include = "./recruiter/nlqs/database" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Flask = "^3.0.3"
pypdf = "^4.2.0"
langchain-openai = "^0.1.13"
langchain-community = "^0.2.6"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"


gradio = "^4.37.2"
chromadb = "^0.5.3"
pandas = "^2.2.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-bot = "recruiter.server:run_server"